requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
decorator==4.4.2
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy==1.19.4
Pillow==8.0.1
pkg-resources==0.0.0
proglog==0.1.9
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
tqdm==4.52.0
urllib3==1.26.2
